[ADD] Stub classes for the podcast pipleine
These are set up to tricle down the information from raw,
to saved, to processed and finally to be downloaded.
In the future it should be able to add more bits in or
rearange the flow.
The first change to make is to make the feed download not
always have to happen at the start.