[ADD] Fleshed out feedScrapper
This now tries to get the xml docs and turn them into a list of items
(each item is a podcast) in dictionary form. At this point there are
a number of things that could have gone wrong with the data prep up
until this point, so a lot more debug and logging needs to be done now.
All the errors should be handled before hand and -at the least- the
correct structures passed down to the later functions for them to
handle correctly. Having trouble testing from feedGrabber without
an internet connection - so I tried to make it use the cached file.
Needs more work.